[HTML-Generator] Simple generator, demonstrating composition and aggregation in Python.
This simple html generator produces .html file. It generates
tags for the body (no nesting, so no div containers and
such), and a title for head and doctype. It then generates
index.html with the defined tags.

Some imporvements can include (but outside of the scope or
purpose of this project):
1) Allows nesting tags.
2) Produce well-organised html document.
3) Allow meta tags and hyperlinks in the head section.